+++ src/utils/print.hpp
@@ -10,14 +10,14 @@
 #include <string>
 
 namespace uw12::print {
-    inline void print_character_line(const char character, const int n) {
-        for (int i = 0; i < n; ++i) {
+    inline void print_character_line(const char character, const size_t n) {
+        for (size_t i = 0; i < n; ++i) {
             std::cout << character;
         }
         std::cout << std::endl;
     }
 
-    inline void print_header(const std::string &string, const int width = 48) {
+    inline void print_header(const std::string &string, const size_t width = 48) {
         std::cout << std::endl;
         print_character_line('=', width);
         std::cout << " " << string << std::endl;
@@ -32,7 +32,7 @@ namespace uw12::print {
 
         const auto pad = width - size;
         std::cout << " " << string << ":";
-        for (int i = 0; i < pad; ++i) {
+        for (size_t i = 0; i < pad; ++i) {
             std::cout << " ";
         }
         std::cout << std::setw(12) << std::fixed << std::right << std::setprecision(6)
